backend/requirements.txt
# EnviroAgent Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & ML Libraries
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
crewai>=0.1.0
openai>=1.3.7
anthropic>=0.7.8

# Machine Learning
torch>=2.2.0
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=2.1.4

# Database & Caching
asyncpg==0.29.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1

# Environment Integrations
requests==2.31.0
httpx==0.25.2
websockets==12.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
